Rigged up a global data context
Added a simple, int only hierarchical data context and integrated it with the FlowAsset

Created a very rough, still unfinished, custom property editor that enables selecting properties from a data context attached to a flowasset without having to type names.

Keys are identified by GUIDs so renaming them will not cause breakage.

There's probably a lot of stuff in there that's not good and needs to be improved. But for the moment it'll do. As little draft. So to speak.