I got caught up in a unity mindset right there. Fixed!
Made it so that the runtime will not affect template game data context by instantiating runtime versions of the data context with current world as outer.

While Unity simply never saves anything at all, unreal tends to keep all the changes. Which is nice. Just need to get used to it.

Later down the line this can be utilised to load game state.